ceph-common: avoid "ceph-mon-all" upstart task
The "ceph-mon-all" task is buggy and does not really reflect the true
state of the monitor processes. It's really only good for ensuring that
every monitor starts when a system first boots. It's not really good at
doing something like "restart every monitor".

Drop the "ceph-mon-all" task and do something similar to what we're
doing with systemd, instead.